reorganize refl1d_interface
